Struck through no sharing
Struck through the sentence on no sharing and updated the last revised date.  It was pointed out that the link to the crossed out Twitter logo was no longer available (404'd) but attempts to find a new image able to be expressly freely used sparked further discussion of whether this entire sentence was moot in a virtual conference with free registration and whose content will be posted. Because the sentence had been part of the Code, it was deemed to use strikethrough rather than remove it so that participants will know what changed. The bad link was removed, however.