Adding Quentin as Convenor of this TDWG IG
Added Quentin as Convenor for this TDWG IG. Left Steve's name as Former Convenor (or we'll lose this information).
Not sure if this IG has a GitHub presence or not. (In other words, are there other places where we need to put Quentin's name and contact information)? I also added two resources.